Spin events off into events.h
